+++ src/LibHac/FsSrv/Sf/IProgramRegistry.cs
@@ -6,8 +6,8 @@ namespace LibHac.FsSrv.Sf;
 
 public interface IProgramRegistry : IDisposable
 {
-    Result RegisterProgram(ulong processId, ProgramId programId, StorageId storageId,
-        InBuffer accessControlData, InBuffer accessControlDescriptor);
+    Result RegisterProgram(ulong processId, ProgramId programId, StorageId storageId, InBuffer accessControlData,
+        long accessControlDataSize, InBuffer accessControlDescriptor, long accessControlDescriptorSize);
 
     Result UnregisterProgram(ulong processId);
     Result SetCurrentProcess(ulong processId);
